Add files via upload
This is the index page of a test BlazorServer project used to provide a sign-up form. The EditForm uses the "@Model" for validation. The custom validation messages will display for individual field errors however the entire form will NOT submit unless it is field in by the user with no errors and in the first attempt. It might require some kind of "State Has Changed" method for the form to refresh as an incorrect entry also throws an error on the whole page.